Refactor infinispan-client extension packages to be consistent with other extensions
