+++ extensions/infinispan-client/deployment/src/main/java/org/jboss/shamrock/infinispan/client/PropertiesBuildItem.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.jboss.shamrock.infinispan;
+package org.jboss.shamrock.infinispan.client;
 
 import java.util.Properties;
