+++ extensions/infinispan-client/runtime/src/main/java/org/jboss/shamrock/infinispan/client/runtime/HandleProtostreamMarshaller.java
@@ -1,4 +1,4 @@
-package org.infinispan.protean.runtime;
+package org.jboss.shamrock.infinispan.client.runtime;
 
 import java.io.IOException;
 import java.io.InputStream;
