+++ extensions/infinispan-client/runtime/src/main/java/org/jboss/shamrock/infinispan/client/runtime/InfinispanConfiguration.java
@@ -1,4 +1,4 @@
-package org.infinispan.protean.runtime;
+package org.jboss.shamrock.infinispan.client.runtime;
 
 import java.util.Optional;
