+++ integration-tests/infinispan-client/src/main/java/org/jboss/shamrock/example/infinispanclient/TestServlet.java
@@ -55,11 +55,11 @@ import org.infinispan.counter.api.CounterConfiguration;
 import org.infinispan.counter.api.CounterManager;
 import org.infinispan.counter.api.CounterType;
 import org.infinispan.counter.api.StrongCounter;
-import org.infinispan.protean.runtime.Remote;
 import org.infinispan.query.api.continuous.ContinuousQuery;
 import org.infinispan.query.api.continuous.ContinuousQueryListener;
 import org.infinispan.query.dsl.Query;
 import org.infinispan.query.dsl.QueryFactory;
+import org.jboss.shamrock.infinispan.client.runtime.Remote;
 import org.jboss.shamrock.runtime.StartupEvent;
 
 @Path("/")
